added Input Box context to prop management flow and further widgetization of a components
