+++ likeminds-chat-reactnative-integration/ChatSX/components/ExploreFeedFilters/index.tsx
@@ -1,46 +1,54 @@
-import React, {useEffect, useRef, useState} from 'react';
+import React, { useState } from "react";
 import {
   View,
   Text,
   Image,
   TouchableOpacity,
   Modal,
-  Alert,
   Pressable,
-} from 'react-native';
-import STYLES from '../../constants/Styles';
-import {useAppDispatch} from '../../store';
-import {getExploreFeedData} from '../../store/actions/explorefeed';
-import {styles} from './styles';
-import {Events, Keys} from '../../enums';
-import {LMChatAnalytics} from '../../analytics/LMChatAnalytics';
+} from "react-native";
+import { styles } from "./styles";
+import { Events, Keys } from "../../enums";
+import { LMChatAnalytics } from "../../analytics/LMChatAnalytics";
+import { useExploreFeedContext } from "../../context/ExploreFeedContext";
 
-interface Props {
-  isPinned: boolean;
-  setIsPinned: (val: any) => void;
-  setFilterState: (val: any) => void;
-  filterState: any;
-  pinnedChatroomsCount: number;
-}
-
-const ExploreFeedFilters = ({
-  isPinned,
-  filterState,
-  setFilterState,
-  setIsPinned,
-  pinnedChatroomsCount,
-}: Props) => {
+const ExploreFeedFilters = () => {
   const [modalVisible, setModalVisible] = useState(false);
+
+  const {
+    chats,
+    pinnedChatroomsCount,
+    filterState,
+    isPinned,
+    exploreChatrooms,
+
+    setIsPinned,
+    setFilterState,
+    setChats,
+  } = useExploreFeedContext();
 
   const handleModalClose = () => {
     setModalVisible(false);
   };
   const arr = [
-    'Newest',
-    'Recently active',
-    'Most messages',
-    'Most participants',
+    "Newest",
+    "Recently active",
+    "Most messages",
+    "Most participants",
   ];
+
+  const handleIsPinned = (val) => {
+    if (val) {
+      const pinnedChats = chats.filter((item: any) =>
+        item?.isPinned ? item : null
+      );
+      setChats(pinnedChats);
+      setIsPinned(val);
+    } else {
+      setChats(exploreChatrooms);
+      setIsPinned(val);
+    }
+  };
 
   return (
     <View>
@@ -49,10 +57,11 @@ const ExploreFeedFilters = ({
           onPress={() => {
             setModalVisible(!modalVisible);
           }}
-          style={styles.itemContainer}>
+          style={styles.itemContainer}
+        >
           <Text style={styles.titleText}>{arr[filterState]}</Text>
           <Image
-            source={require('../../assets/images/down_arrow3x.png')}
+            source={require("../../assets/images/down_arrow3x.png")}
             style={styles.icon}
           />
         </TouchableOpacity>
@@ -60,25 +69,27 @@ const ExploreFeedFilters = ({
         {isPinned ? (
           <TouchableOpacity
             onPress={() => {
-              setIsPinned(false);
+              handleIsPinned(false);
             }}
-            style={styles.cancelPinnedBtn}>
+            style={styles.cancelPinnedBtn}
+          >
             <View style={styles.cancelPinIconParent}>
               <Image
-                source={require('../../assets/images/pin_icon_blue3x.png')}
+                source={require("../../assets/images/pin_icon_blue3x.png")}
                 style={styles.cancelPinIcon}
               />
             </View>
             <View
               style={{
-                display: 'flex',
-                flexDirection: 'row',
-                alignItems: 'center',
-                justifyContent: 'center',
-              }}>
+                display: "flex",
+                flexDirection: "row",
+                alignItems: "center",
+                justifyContent: "center",
+              }}
+            >
               <Text style={styles.titleText}>Pinned</Text>
               <Image
-                source={require('../../assets/images/cross_icon3x.png')}
+                source={require("../../assets/images/cross_icon3x.png")}
                 style={styles.icon}
               />
             </View>
@@ -89,23 +100,24 @@ const ExploreFeedFilters = ({
               setIsPinned(true);
               LMChatAnalytics.track(
                 Events.PINNED_CHATROOM_VIEWED,
-                new Map<string, string>([[Keys.SOURCE, 'overflow_menu']]),
+                new Map<string, string>([[Keys.SOURCE, "overflow_menu"]])
               );
-            }}>
+            }}
+          >
             <Image
-              source={require('../../assets/images/pin_icon_grey3x.png')}
+              source={require("../../assets/images/pin_icon_grey3x.png")}
               style={styles.pinIcon}
             />
           </TouchableOpacity>
         ) : null}
       </View>
       <Modal
-        // animationType="slide"
         transparent={true}
         visible={modalVisible}
         onRequestClose={() => {
           setModalVisible(!modalVisible);
-        }}>
+        }}
+      >
         <Pressable style={styles.centeredView} onPress={handleModalClose}>
           <View>
             <Pressable onPress={() => {}} style={[styles.modalView]}>
@@ -115,7 +127,8 @@ const ExploreFeedFilters = ({
                     setFilterState(index);
                   }}
                   key={val + index}
-                  style={styles.filtersView}>
+                  style={styles.filtersView}
+                >
                   <Text style={styles.filterText}>{val}</Text>
                 </TouchableOpacity>
               ))}
