+++ likeminds-chat-reactnative-integration/ChatSX/components/MessageList/index.tsx
@@ -1,18 +1,12 @@
-import { View, Text, Pressable, Image } from "react-native";
+import { View, Text, Pressable } from "react-native";
 import React from "react";
 import { FlashList } from "@shopify/flash-list";
 import Swipeable from "../Swipeable";
 import Messages from "../Messages";
-import { decode } from "../../commonFuctions";
-import { ChatroomType, DocumentType } from "../../enums";
 import { useAppDispatch, useAppSelector } from "../../store";
 import STYLES from "../../constants/Styles";
-import {
-  GET_CONVERSATIONS_SUCCESS,
-  SET_POSITION,
-} from "../../store/types/types";
+import { SET_POSITION } from "../../store/types/types";
 import { styles } from "./styles";
-import { getCurrentConversation } from "../../utils/chatroomUtils";
 import { useLMChatStyles } from "../../lmChatProvider";
 import {
   ChatroomContextValues,
@@ -22,8 +16,10 @@ import LinearGradient from "react-native-linear-gradient";
 import { createShimmerPlaceholder } from "react-native-shimmer-placeholder";
 import {
   MessageListContextProvider,
+  MessageListContextValues,
   useMessageListContext,
 } from "../../context/MessageListContext";
+import ChatroomTopic from "../ChatroomTopic";
 
 const ShimmerPlaceHolder = createShimmerPlaceholder(LinearGradient);
 
@@ -38,13 +34,9 @@ const MessageList = () => {
 const MessageListComponent = () => {
   const {
     conversations,
-    chatroomID,
-    user,
     selectedMessages,
     currentChatroomTopic,
-    chatroomType,
     shimmerIsLoading,
-    chatroomName,
     refInput,
 
     handleLongPress,
@@ -54,14 +46,9 @@ const MessageListComponent = () => {
   const {
     flatlistRef,
     isFound,
-    flashListMounted,
-
-    setIsFound,
-    setFlashListMounted,
     handleOnScroll,
     renderFooter,
-    getIconAttachment,
-  } = useMessageListContext();
+  }: MessageListContextValues = useMessageListContext();
 
   const LMChatContextStyles = useLMChatStyles();
   const chatBubbleStyles = LMChatContextStyles?.chatBubbleStyles;
@@ -82,6 +69,7 @@ const MessageListComponent = () => {
 
   return (
     <>
+      {/* loading shimmer */}
       {shimmerIsLoading ? (
         <View style={{ marginTop: 10 }}>
           <View
@@ -187,6 +175,7 @@ const MessageListComponent = () => {
         </View>
       ) : (
         <>
+          {/* List of messages */}
           <FlashList
             ref={flatlistRef}
             data={conversations}
@@ -287,7 +276,12 @@ const MessageListComponent = () => {
                           : null
                       }
                     >
-                      <Messages isIncluded={isIncluded} item={item} />
+                      <Messages
+                        isIncluded={isIncluded}
+                        item={item}
+                        isStateIncluded={isStateIncluded}
+                        index={index}
+                      />
                     </Pressable>
                   </Swipeable>
                 </View>
@@ -301,138 +295,7 @@ const MessageListComponent = () => {
           />
 
           {/* Chatroom Topic */}
-          {!(Object.keys(currentChatroomTopic).length === 0) &&
-          chatroomType !== ChatroomType.DMCHATROOM ? (
-            <Pressable
-              style={styles.chatroomTop}
-              onPress={async () => {
-                const index = conversations.findIndex(
-                  (element: any) => element?.id == currentChatroomTopic?.id
-                );
-                if (index >= 0) {
-                  if (!flashListMounted) {
-                    setTimeout(() => {
-                      flatlistRef.current?.scrollToIndex({
-                        animated: true,
-                        index,
-                      });
-                      setIsFound(true);
-                      setFlashListMounted(true);
-                    }, 1000);
-                  } else {
-                    flatlistRef.current?.scrollToIndex({
-                      animated: true,
-                      index,
-                    });
-                    setIsFound(true);
-                  }
-                } else {
-                  const newConversation = await getCurrentConversation(
-                    currentChatroomTopic,
-                    chatroomID?.toString()
-                  );
-                  dispatch({
-                    type: GET_CONVERSATIONS_SUCCESS,
-                    body: { conversations: newConversation },
-                  });
-                  const index = newConversation.findIndex(
-                    (element) => element?.id == currentChatroomTopic?.id
-                  );
-                  if (index >= 0) {
-                    flatlistRef.current?.scrollToIndex({
-                      animated: true,
-                      index,
-                    });
-                    setIsFound(true);
-                  }
-                }
-              }}
-            >
-              <View style={styles.chatroomTopic}>
-                <View>
-                  <Image
-                    source={
-                      currentChatroomTopic?.member?.imageUrl
-                        ? { uri: currentChatroomTopic?.member?.imageUrl }
-                        : require("../../assets/images/default_pic.png")
-                    }
-                    style={styles.chatroomTopicAvatar}
-                  />
-                </View>
-                <View style={styles.chatRoomTopicInfo}>
-                  <Text
-                    ellipsizeMode="tail"
-                    numberOfLines={1}
-                    style={{
-                      color: STYLES.$COLORS.FONT_PRIMARY,
-                      fontSize: STYLES.$FONT_SIZES.LARGE,
-                      fontFamily: STYLES.$FONT_TYPES.BOLD,
-                    }}
-                  >
-                    Current Topic
-                  </Text>
-
-                  {currentChatroomTopic?.hasFiles == true ? (
-                    getIconAttachment(currentChatroomTopic)
-                  ) : currentChatroomTopic?.state === 10 ? (
-                    getIconAttachment(currentChatroomTopic)
-                  ) : currentChatroomTopic?.ogTags?.url !== null &&
-                    currentChatroomTopic?.ogTags ? (
-                    getIconAttachment(currentChatroomTopic)
-                  ) : (
-                    <Text
-                      ellipsizeMode="tail"
-                      numberOfLines={2}
-                      style={{
-                        color: STYLES.$COLORS.MSG,
-                        fontSize: STYLES.$FONT_SIZES.MEDIUM,
-                        fontFamily: STYLES.$FONT_TYPES.LIGHT,
-                        lineHeight: 18,
-                      }}
-                    >
-                      {decode({
-                        text: currentChatroomTopic?.answer,
-                        enableClick: false,
-                        chatroomName: chatroomName,
-                        communityId: user?.sdkClientInfo?.community,
-                      })}
-                    </Text>
-                  )}
-                </View>
-                <View>
-                  {currentChatroomTopic?.attachmentCount > 0 &&
-                  currentChatroomTopic?.attachments.length > 0 &&
-                  currentChatroomTopic?.attachments[0]?.type !==
-                    DocumentType.PDF &&
-                  currentChatroomTopic?.attachments[0]?.type !==
-                    DocumentType.VOICE_NOTE ? (
-                    <Image
-                      source={{
-                        uri:
-                          currentChatroomTopic?.attachments[0]?.type ==
-                          DocumentType.IMAGE
-                            ? currentChatroomTopic?.attachments[0]?.url
-                            : currentChatroomTopic?.attachments[0]
-                                ?.thumbnailUrl,
-                      }}
-                      style={styles.chatroomTopicAttachment}
-                    />
-                  ) : currentChatroomTopic?.ogTags?.url !== null &&
-                    currentChatroomTopic?.ogTags?.image !== "" &&
-                    currentChatroomTopic?.ogTags !== undefined &&
-                    currentChatroomTopic?.ogTags !== null &&
-                    Object.keys(currentChatroomTopic).length !== 0 ? (
-                    <Image
-                      source={{
-                        uri: currentChatroomTopic?.ogTags?.image,
-                      }}
-                      style={styles.chatroomTopicAttachment}
-                    />
-                  ) : null}
-                </View>
-              </View>
-            </Pressable>
-          ) : null}
+          <ChatroomTopic />
         </>
       )}
     </>
