+++ likeminds-chat-reactnative-integration/ChatSX/customModals/ApproveDMRequest/index.tsx
@@ -1,24 +1,31 @@
-import {View, Text, Modal, Pressable, TouchableOpacity} from 'react-native';
-import React from 'react';
-import {styles} from '../styles';
+import { View, Text, Modal, Pressable, TouchableOpacity } from "react-native";
+import React from "react";
+import { styles } from "../styles";
 import {
   APPROVE_BUTTON,
   APPROVE_DM_REQUEST,
   APPROVE_REQUEST_MESSAGE,
   CANCEL_BUTTON,
-} from '../../constants/Strings';
+} from "../../constants/Strings";
+import {
+  ChatroomContextValues,
+  useChatroomContext,
+} from "../../context/ChatroomContext";
 
-const ApproveDMRequestModal = ({
-  hideDMApproveAlert,
-  DMApproveAlertModalVisible,
-  onApprove,
-}: any) => {
+const ApproveDMRequestModal = () => {
+  const {
+    DMApproveAlertModalVisible,
+    
+    onApprove,
+    hideDMApproveAlert,
+  }: ChatroomContextValues = useChatroomContext();
   return (
     <Modal
       visible={DMApproveAlertModalVisible}
       animationType="fade"
       transparent={true}
-      onRequestClose={hideDMApproveAlert}>
+      onRequestClose={hideDMApproveAlert}
+    >
       <Pressable style={styles.modal} onPress={hideDMApproveAlert}>
         <Pressable onPress={() => {}} style={styles.modalContainer}>
           <Text style={styles.title}>{APPROVE_DM_REQUEST}</Text>
@@ -26,7 +33,8 @@ const ApproveDMRequestModal = ({
           <View style={styles.buttonContainer}>
             <TouchableOpacity
               style={[styles.button, styles.cancelButton]}
-              onPress={hideDMApproveAlert}>
+              onPress={hideDMApproveAlert}
+            >
               <Text style={[styles.buttonText, styles.cancelButtonText]}>
                 {CANCEL_BUTTON}
               </Text>
@@ -36,7 +44,8 @@ const ApproveDMRequestModal = ({
               onPress={() => {
                 onApprove();
                 hideDMApproveAlert();
-              }}>
+              }}
+            >
               <Text style={styles.buttonText}>{APPROVE_BUTTON}</Text>
             </TouchableOpacity>
           </View>
