+++ likeminds-chat-reactnative-integration/ChatSX/customModals/WarningMessage/index.tsx
@@ -1,27 +1,46 @@
-import {View, Text, Modal, Pressable, TouchableOpacity} from 'react-native';
-import React from 'react';
-import {styles} from '../styles';
-import {CONFIRM_BUTTON, CANCEL_BUTTON} from '../../constants/Strings';
+import { View, Text, Modal, Pressable, TouchableOpacity } from "react-native";
+import React from "react";
+import { styles } from "../styles";
+import {
+  CONFIRM_BUTTON,
+  CANCEL_BUTTON,
+  WARNING_MSG_PRIVATE_CHATROOM,
+  WARNING_MSG_PUBLIC_CHATROOM,
+} from "../../constants/Strings";
+import {
+  ChatroomContextValues,
+  useChatroomContext,
+} from "../../context/ChatroomContext";
 
-const WarningMessageModal = ({
-  hideWarningModal,
-  warningMessageModalState,
-  warningMessage,
-  leaveChatroom,
-}: any) => {
+const WarningMessageModal = () => {
+  const {
+    isWarningMessageModalState,
+    isSecret,
+
+    leaveChatroom,
+    leaveSecretChatroom,
+    hideWarningModal,
+  }: ChatroomContextValues = useChatroomContext();
+
+  const warningMessage = isSecret
+    ? WARNING_MSG_PRIVATE_CHATROOM
+    : WARNING_MSG_PUBLIC_CHATROOM;
+
   return (
     <Modal
-      visible={warningMessageModalState}
+      visible={isWarningMessageModalState}
       animationType="fade"
       transparent={true}
-      onRequestClose={hideWarningModal}>
+      onRequestClose={hideWarningModal}
+    >
       <Pressable style={styles.modal} onPress={hideWarningModal}>
         <Pressable onPress={() => {}} style={styles.modalContainer}>
           <Text style={styles.message}>{warningMessage}</Text>
           <View style={styles.buttonContainer}>
             <TouchableOpacity
               style={[styles.button, styles.cancelButton]}
-              onPress={hideWarningModal}>
+              onPress={hideWarningModal}
+            >
               <Text style={[styles.buttonText, styles.cancelButtonText]}>
                 {CANCEL_BUTTON}
               </Text>
@@ -29,9 +48,16 @@ const WarningMessageModal = ({
             <TouchableOpacity
               style={[styles.button, styles.okButton]}
               onPress={() => {
-                leaveChatroom();
+                () => {
+                  if (isSecret) {
+                    leaveSecretChatroom();
+                  } else {
+                    leaveChatroom();
+                  }
+                };
                 hideWarningModal();
-              }}>
+              }}
+            >
               <Text style={styles.buttonText}>{CONFIRM_BUTTON}</Text>
             </TouchableOpacity>
           </View>
