+++ likeminds-chat-reactnative-integration/ChatSX/shareUtils/index.ts
@@ -1,9 +1,9 @@
-import {Alert, Share} from 'react-native';
-import {VALID_URI_REGEX} from '../constants/Regex';
-import {ShareChatroomRequest} from './models';
-import {Events, Keys} from '../enums';
-import {LMChatAnalytics} from '../analytics/LMChatAnalytics';
-import {getChatroomType} from '../utils/analyticsUtils';
+import { Alert, Share } from "react-native";
+import { VALID_URI_REGEX } from "../constants/Regex";
+import { ShareChatroomRequest } from "./models";
+import { Events, Keys } from "../enums";
+import { LMChatAnalytics } from "../analytics/LMChatAnalytics";
+import { getChatroomType } from "../utils/analyticsUtils";
 
 // this method generates URL for share
 export const shareChatroomURL = ({
@@ -22,26 +22,26 @@ export function isValidURI(uri: string): boolean {
 
 // this function helps to share chatroom url
 export const onShare = async (
-  chatroomID: number,
+  chatroomID: string,
   chatroomType: number,
-  isSecret?: boolean,
+  isSecret?: boolean
 ) => {
   try {
     const shareChatroomRequest = {
       chatroomId: chatroomID,
-      domain: '', // Add your custom link to open app,
+      domain: "", // Add your custom link to open app,
     };
     const shareUrl = shareChatroomURL(shareChatroomRequest);
     LMChatAnalytics.track(
       Events.CHAT_ROOM_SHARED,
       new Map<string, string>([
         [Keys.SOURCE, chatroomID?.toString()],
-        [Keys.CHATROOM_FEED, 'chatroom'],
+        [Keys.CHATROOM_FEED, "chatroom"],
         [
           Keys.CHATROOM_TYPE,
           getChatroomType(chatroomType?.toString(), isSecret),
         ],
-      ]),
+      ])
     );
     await Share.share({
       message: shareUrl,
